+++ packages/atlas/src/views/legal/PrivacyPolicy.tsx
@@ -4,12 +4,12 @@ import { MarkdownPreview } from '@/components/MarkdownPreview'
 import { atlasConfig } from '@/config'
 import { useHeadTags } from '@/hooks/useHeadTags'
 
-export const CopyrightPolicyView: FC = () => {
-  const headTags = useHeadTags('Copyright Policy')
+export const PrivacyPolicyView: FC = () => {
+  const headTags = useHeadTags('Privacy Policy')
   return (
     <>
       {headTags}
-      <MarkdownPreview markdown={atlasConfig.legal.copyrightPolicy} />
+      <MarkdownPreview markdown={atlasConfig.legal.privacyPolicy} />
     </>
   )
 }