+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Stop video autoplay if user starts commenting
 - Fix mobile touch scroll on bottom drawers
+- Expose Privacy Policy in config and introduce basic Markdown support
 
 ## [1.1.0] - 2022-11-30